Stop using Flow as the public API for pulses
Using `Flow` as the API surface for scheduling recurring jobs
is potentially problematic. The library of operators on `Flow`
can potentially interfere with expectation of value emissions.
`Pulse` instances are backpressure sensitive, and therefore
we should start with a restrictive API that prevent consumers
from applying misguided backpressure.

Closes #5
