+++ src/commonTest/kotlin/io/github/kevincianfarini/cardiologist/PulseTests.kt
@@ -1,29 +1,21 @@
 package io.github.kevincianfarini.cardiologist
 
 import app.cash.turbine.test
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.merge
 import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.test.testTimeSource
-import kotlinx.coroutines.withContext
 import kotlinx.datetime.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertTrue
-import kotlin.time.Duration.Companion.days
-import kotlin.time.Duration.Companion.milliseconds
-import kotlin.time.Duration.Companion.nanoseconds
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.ExperimentalTime
 import kotlin.time.measureTime
 
 @OptIn(ExperimentalTime::class, ExperimentalCoroutinesApi::class)
-class FlowTests {
+class PulseTests {
 
     @Test fun durationPulse_emits_initial_instant_before_delaying() = runTest {
-        testClock.intervalPulse(5.seconds).test {
+        testClock.intervalPulse(5.seconds).flow.test {
             assertEquals(
                 expected = 0.seconds,
                 actual = testTimeSource.measureTime { awaitItem() }
@@ -32,7 +24,7 @@ class FlowTests {
     }
 
     @Test fun durationPulse_emits_subsequent_instants_after_delaying() = runTest {
-        testClock.intervalPulse(5.seconds).test {
+        testClock.intervalPulse(5.seconds).flow.test {
             skipItems(1)
             assertEquals(
                 expected = 5.seconds,
@@ -42,7 +34,7 @@ class FlowTests {
     }
 
     @Test fun periodPulse_emits_initial_instant_before_delaying() = runTest {
-        testClock.intervalPulse(DateTimePeriod(seconds = 5), TimeZone.UTC).test {
+        testClock.intervalPulse(DateTimePeriod(seconds = 5), TimeZone.UTC).flow.test {
             assertEquals(
                 expected = 0.seconds,
                 actual = testTimeSource.measureTime { awaitItem() }
@@ -51,7 +43,7 @@ class FlowTests {
     }
 
     @Test fun periodPulse_emits_subsequent_instants_after_delaying() = runTest {
-        testClock.intervalPulse(DateTimePeriod(seconds = 5), TimeZone.UTC).test {
+        testClock.intervalPulse(DateTimePeriod(seconds = 5), TimeZone.UTC).flow.test {
             skipItems(1)
             assertEquals(
                 expected = 5.seconds,
